requirements.txt
# Streamlit Cloud dependencies
# Generated from pyproject.toml for deployment

duckdb>=1.1.0
pandas>=2.2.0
geopandas>=1.0.0
networkx>=3.2.0
gtfs-kit>=12.0
httpx>=0.28.0
python-dotenv>=1.0.0
loguru>=0.7.0
pooch>=1.8.0
numpy

# Visualization
streamlit>=1.40.0
keplergl>=0.3.6
streamlit-keplergl>=0.3.0
matplotlib>=3.9.0
